remove @expo/fastlane lib
