+++ packages/build-tools/package.json
@@ -23,7 +23,6 @@
     "@expo/config-plugins": "4.1.0",
     "@expo/downloader": "0.0.19",
     "@expo/eas-build-job": "0.2.68",
-    "@expo/fastlane": "0.0.28",
     "@expo/logger": "0.0.22",
     "@expo/package-manager": "^0.0.47",
     "@expo/plist": "^0.0.11",
