--- packages/fastlane/README.md
@@ -1,7 +1,0 @@
-# @expo/fastlane
-
-`@expo/fastlane` is a light wrapper that calls fastlane executable in your PATH.
-
-## Repository
-
-https://github.com/expo/eas-build/tree/main/src/packages/fastlane
