--- packages/fastlane/jest/setup-tests.ts
@@ -1,5 +1,0 @@
-if (process.env.NODE_ENV !== 'test') {
-  throw new Error("NODE_ENV environment variable must be set to 'test'.");
-}
-
-// Always mock:
