--- packages/fastlane/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@expo/fastlane",
-  "version": "0.0.28",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "start": "yarn watch",
-    "watch": "tsc --watch --preserveWatchOutput",
-    "build": "tsc",
-    "prepack": "rm -rf dist && tsc -p tsconfig.build.json",
-    "test": "jest --config jest.config.js --passWithNoTests",
-    "test:watch": "jest --config jest.config.js --watch"
-  },
-  "author": "Expo <support@expo.io>",
-  "bugs": "https://github.com/expo/eas-build/issues",
-  "license": "BUSL-1.1",
-  "dependencies": {
-    "@expo/logger": "0.0.22",
-    "@expo/turtle-spawn": "0.0.23"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.20",
-    "@types/node": "^16.11.26",
-    "jest": "^27.5.1",
-    "ts-jest": "^27.1.3",
-    "typescript": "^4.6.2"
-  }
-}
