--- packages/fastlane/tsconfig.build.json
@@ -1,4 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": ["src/**/__mocks__/**/*.ts", "src/**/__tests__/**/*.ts"]
-}
